Expose cutom extension issue
